requirements/base.txt
Django==3.2.25
  sqlparse==0.5.0
  pytz==2023.3
dealer==2.0.5
django-crispy-forms==1.12.0
django-dotenv==1.4.2 # To manage environment variables
django-formtools==2.2 # For multi-page forms
psycopg2-binary==2.9.5 # Database backend
Pillow==10.3.0
python3-saml==1.9.0 # For SSO support
  tablib==0.14.0  # Note: version 1.0.0 seems to be broken
  xmlsec==1.3.13 # 1.3.14 is not compatible with python3-saml==1.9.0
requests==2.32.3
pandas==1.5.3
django-q2==1.5.5 # For async tasks
  croniter==1.3.5 
  psutil==5.9.5
# Until Pango for Centos is upgraded to 1.44 we can't upgrade weasyprint to latest
weasyprint==52.5 # For generating PDF assay reports
factory_boy==2.11.1 # For testing models
gunicorn==22.0.0
django-storages==1.14.2 # For AWS file storage
boto3==1.34.4 # For AWS logging
watchtower==3.0.0 # For AWS logging
sentry-sdk==2.8.0 # For Sentry logging

# For development
ipython==8.25.0
django-debug-toolbar==3.2.1
coverage==4.5.4
flake8==3.7.8
  entrypoints==0.3
  pycodestyle==2.5.0
  pyflakes==2.1.1
djhtml==v1.4.10
isort==5.6.3
black==22.3.0
debugpy==1.5.1
